backend/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aircraftworth"
version = "0.1.0"
description = "MLAT aircraft tracking system using Neuron sensors and Hedera consensus"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "AircraftWorth Team"}
]
keywords = ["mlat", "aircraft", "tracking", "hedera", "neuron", "adsb"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: GIS",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "hedera-sdk-py>=2.50.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "httpx>=0.27.0",
    "supabase>=2.4.1",
    "numpy>=1.26.4",
    "scipy>=1.11.4",
    "pyproj>=3.6.1",
    "pyModeS>=2.15",
    "aiohttp>=3.9.1",
    "groq>=0.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.21.1",
]
groq = [
    "groq>=0.4.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = [
    "mlat_core*",
    "hedera_logger*",
    "neuron_client*",
    "services*",
    "routers*",
    "models*",
]
exclude = [
    "node_modules*",
    "supabase*",
    "scripts*",
    "tests*",
    "htmlcov*",
    "__pycache__*",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=. --cov-report=html --cov-report=term-missing --cov-fail-under=100"
